27/12/2023
I found out there was a dump for the full version of Sonic 1, with special stages and sound effects. I can't seem to find where the graphics and audio data is though. But there are 2 class files and more level files, so I'll leave it in it's own folder for now.